fix(testing): slim quarantine CI script and harden legacy slug regex
Delegate ledger validation/output to TypeScript scripts to remove
duplication with quarantine-ledger.ts and avoid ReDoS-prone regex.
